+++ src/cookie/cookie.go
@@ -1,13 +1,10 @@
-package server
+package cookie
 
 import (
-	"encoding/base64"
-	"github.com/gorilla/websocket"
 	"github.com/gorilla/sessions"
 	"log"
 	"net/http"
 	"strconv"
-	"webtty"
 	"utils"
 	"user"
 	"errors"
@@ -55,7 +52,7 @@ func DecrementCounterCookies(rw http.ResponseWriter, req *http.Request) {
 	}
 }
 
-func GetCounterCookieValue(rw http.ResponseWriter, req *http.Request) int {
+func GetCounterCookieValue(req *http.Request) int {
 	cookie, err := req.Cookie("Session-Counter")
 	if err != nil {
 		log.Println("Error while getting cookie: ", err.Error())
@@ -65,17 +62,10 @@ func GetCounterCookieValue(rw http.ResponseWriter, req *http.Request) int {
 	return sessionCount
 }
 
-func WriteMessageToTerminal(conn *websocket.Conn, message string) {
-	safeMessage := base64.StdEncoding.EncodeToString([]byte(message))
-	err := conn.WriteMessage(websocket.TextMessage, []byte(append([]byte{webtty.Output}, []byte(safeMessage)...)))
-	if err != nil {
-		log.Println("err while writing: ", err)
-	}
-}
-
 
 var session_store *sessions.CookieStore
 
+// initilize the cookiestore with secret stored in session DB
 func Init_SessionStore(secret string) {
 	session_store = sessions.NewCookieStore([]byte(secret))
 }
@@ -143,7 +133,7 @@ func Delete_SessionCookie(rw http.ResponseWriter, req *http.Request, session use
 }
 
 
-func Is_UserLoggedIn(rw http.ResponseWriter, req *http.Request) (loggedin bool) {
+func Is_UserLoggedIn(req *http.Request) (loggedin bool) {
 	session_cookie, err := session_store.Get(req, "user-session")
 	val := session_cookie.Values["loggedIn"]
 	loggedin, ok := val.(bool);
@@ -154,7 +144,7 @@ func Is_UserLoggedIn(rw http.ResponseWriter, req *http.Request) (loggedin bool) 
 	return loggedin
 }
 
-func Get_Uid(rw http.ResponseWriter, req *http.Request) (uid string) {
+func Get_Uid(req *http.Request) (uid string) {
 	session_cookie, _ := session_store.Get(req, "user-session")
 	val := session_cookie.Values["uid"]
 	uid, ok := val.(string);
@@ -164,7 +154,7 @@ func Get_Uid(rw http.ResponseWriter, req *http.Request) (uid string) {
 	return uid
 }
 
-func Get_SessionID(rw http.ResponseWriter, req *http.Request) (sessionid string) {
+func Get_SessionID(req *http.Request) (sessionid string) {
 	session_cookie, _ := session_store.Get(req, "user-session")
 	val := session_cookie.Values["sessionID"]
 	sessionid, ok := val.(string);
@@ -174,7 +164,7 @@ func Get_SessionID(rw http.ResponseWriter, req *http.Request) (sessionid string)
 	return sessionid
 }
 
-func Get_ExpirationTime(rw http.ResponseWriter, req *http.Request) (e int64) {
+func Get_ExpirationTime(req *http.Request) (e int64) {
 	session_cookie, _ := session_store.Get(req, "user-session")
 	val := session_cookie.Values["expirationTime"]
 	e, ok := val.(int64);
@@ -184,11 +174,11 @@ func Get_ExpirationTime(rw http.ResponseWriter, req *http.Request) (e int64) {
 	return e
 }
 
-func Get_SessionCookie(rw http.ResponseWriter, req *http.Request) (session user.UserSession) {
-		session.Uid = Get_Uid(rw, req)
-		session.SessionID = Get_SessionID(rw, req)
-		session.LoggedIn = Is_UserLoggedIn(rw, req)
-		session.ExpirationTime = Get_ExpirationTime(rw, req)
+func Get_SessionCookie(req *http.Request) (session user.UserSession) {
+		session.Uid = Get_Uid(req)
+		session.SessionID = Get_SessionID(req)
+		session.LoggedIn = Is_UserLoggedIn(req)
+		session.ExpirationTime = Get_ExpirationTime(req)
 		return session
 }
